// src/api/generated/model/updateSessionStateRequestSessionState.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export type UpdateSessionStateRequestSessionState =
  (typeof UpdateSessionStateRequestSessionState)[keyof typeof UpdateSessionStateRequestSessionState];

export const UpdateSessionStateRequestSessionState = {
  NOT_STARTED: "NOT_STARTED",
  IN_PROGRESS: "IN_PROGRESS",
  FINISHED: "FINISHED",
  PAUSED: "PAUSED",
  CANCELLED: "CANCELLED",
} as const;